+++ frontend/src/screens/SecurityScreen.js
@@ -1,0 +1,108 @@
+import {Button, Row, Form, Col} from "react-bootstrap";
+import {useEffect, useState} from "react";
+import {useDispatch, useSelector} from "react-redux";
+import {USER_UPDATE_PROFILE_RESET} from "../constants/userConstants";
+import {getUserDetails, updateUserProfile} from "../actions/userActions";
+import Message from "../components/Message";
+import Loader from "../components/Loader";
+
+const SecurityScreen = ({ location, history }) => {
+    const [name, setName] = useState('')
+    const [email, setEmail] = useState('')
+    const [password, setPassword] = useState('')
+    const [confirmPassword, setConfirmPassword] = useState('')
+    const [message, setMessage] = useState(null)
+
+    const dispatch = useDispatch()
+
+    const userDetails = useSelector((state) => state.userDetails)
+    const { loading, error, user } = userDetails
+
+    const userLogin = useSelector((state) => state.userLogin)
+    const { userInfo } = userLogin
+
+    const userUpdateProfile = useSelector((state) => state.userUpdateProfile)
+    const { success } = userUpdateProfile
+
+    useEffect(() => {
+        if (!userInfo) {
+            history.push('/login')
+        } else {
+            if (!user || !user.name || success) {
+                dispatch({ type: USER_UPDATE_PROFILE_RESET })
+                dispatch(getUserDetails('profile'))
+            } else {
+                setName(user.name)
+                setEmail(user.email)
+            }
+        }
+    }, [dispatch, history, userInfo, user, success])
+
+    const submitHandler = (e) => {
+        e.preventDefault()
+        if (password !== confirmPassword) {
+            setMessage('Passwords do not match')
+        } else {
+            dispatch(updateUserProfile({ id: user._id, name, email, password }))
+        }
+    }
+
+    return (
+        <Row>
+            <Col>
+                <h2>Security</h2>
+                {message && <Message variant='danger'>{message}</Message>}
+                {error && <Message variant="danger">{error}</Message>}
+                {success && <Message variant="success">Profile Updated</Message>}
+                {loading && <Loader />}
+                <Form onSubmit={submitHandler}>
+                    <Form.Group controlId="name">
+                        <Form.Label>Name</Form.Label>
+                        <Form.Control
+                            type="name"
+                            placeholder="Enter name"
+                            value={name}
+                            onChange={(e) => setName(e.target.value)}
+                        ></Form.Control>
+                    </Form.Group>
+
+                    <Form.Group className="my-2" controlId="email">
+                        <Form.Label>Email Address</Form.Label>
+                        <Form.Control
+                            type="email"
+                            placeholder="Enter email"
+                            value={email}
+                            onChange={(e) => setEmail(e.target.value)}
+                        ></Form.Control>
+                    </Form.Group>
+
+                    <Form.Group controlId="password">
+                        <Form.Label>Password</Form.Label>
+                        <Form.Control
+                            type="password"
+                            placeholder="Enter password"
+                            value={password}
+                            onChange={(e) => setPassword(e.target.value)}
+                        ></Form.Control>
+                    </Form.Group>
+
+                    <Form.Group className="my-3" controlId="confirmPassword">
+                        <Form.Label>Confirm Password</Form.Label>
+                        <Form.Control
+                            type="password"
+                            placeholder="Confirm password"
+                            value={confirmPassword}
+                            onChange={(e) => setConfirmPassword(e.target.value)}
+                        ></Form.Control>
+                    </Form.Group>
+
+                    <Button className="mb-4" type="submit" variant="primary">
+                        Update
+                    </Button>
+                </Form>
+            </Col>
+        </Row>
+    )
+}
+
+export default SecurityScreen
